Complete the DNS entry manager.
The entry manager, manager.py pulls current leases from Kea, loads
existing zone entries from Unbound config files, and looks for any
DNS entries that are not in the leases.  If differences are found,
the manager overwrites the existing Unbound config file.  If no
differences are found the file is left as-is.  There is definitely
some clean up and simplification that can be made in the manager
moving forward, but this works.

The test_lease_loader can be used to seed Kea with test data and
delete it at the end.  The test_zones.conf can be used for initial
seeding of the zones.conf to create diffs.