Configure zone .local to be type static.
For v1.3 CoreDNS needs .local entries to all DNS.  This mimics
existing /etc/hosts behavior.  By default, unconfigured zones
are of type "transparent" which in this case causes long delays
in CoreDNS lookups.  A zone type of static returns NXDOMAIN
quickly for records not found in the local-data lists.  This
avoids the CoreDNS timeouts.